rename AlbumMetaList to MetaList
